CI - fix running github actions in 2025
https://github.com/EdgedesignCZ/phpqa/actions/runs/15660110579/job/44116580745?pr=265
This is a scheduled Ubuntu 20.04 retirement. Ubuntu 20.04 LTS runner will be removed on 2025-04-15.

https://github.com/EdgedesignCZ/phpqa/actions/runs/15660147405/job/44116674050?pr=265
This request has been automatically failed because it uses a deprecated version of `actions/cache: v2`.
This request has been automatically failed because it uses a deprecated version of `actions/upload-artifact: v2`

https://github.com/EdgedesignCZ/phpqa/actions/runs/15660192541/job/44116788349?pr=265#step:7:170
Ignore security-checker errors in deprecated php versions